agents/simple_langchain_deepagent/pyproject.toml
[project]
name = "simple-langchain-deepagent"
version = "0.1.0"
description = "LangGraph ReAct deep agent with worker service"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3",
    "langgraph>=0.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2",
    "pydantic-settings>=2.0",
    "python-json-logger>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "httpx[asyncio]>=0.27",
    "gitpython>=3.1",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "aiosqlite>=0.20",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "smoke: requires a live Ollama instance",
    "unit: pure unit tests, no I/O",
    "integration: requires compiled graph",
]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["N802", "N803"]
